+++ app/src/main/java/eu/kanade/tachiyomi/data/preference/PreferencesHelper.kt
@@ -406,6 +406,8 @@ class PreferencesHelper(val context: Context, val preferenceStore: PreferenceSto
 
     fun hideInLibraryItems() = preferenceStore.getBoolean("browse_hide_in_library_items", false)
 
+    fun showDuplicateInLibraryItems() = preferenceStore.getBoolean("browse_show_duplicate_in_library_items", false)
+
     // Tutorial preferences
     fun shownFilterTutorial() = preferenceStore.getBoolean("shown_filter_tutorial", false)